pyproject.toml
[tool.poetry]
name = "airikr"
version = "0.1.0"
description = ""
authors = ["Erik Aulin <erik@aulin.co>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
llama-index = "^0.10.22"
llama-index-llms-openai = "^0.1.12"
llama-index-llms-ollama = "^0.1.2"
llama-index-llms-anthropic = "^0.1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
